Use alpine package for netdata
With this, I can run netdata in local mode and see a decent dashboard
that seems to pretty closely match the one from the upstream version.
One difference is the presence of the "Nodes (beta)" button; another
is the absence of the "Sign in" button.

There are a few TODOs marked in the Dockerfile; see the comments in
there for more details:

- Figure out what's going wrong when I copy in netdata.conf

- Copying in balenad.conf *may* have stopped the container graphs from
showing up

- File a bug upstream about creation of /var/lib/netdata

Also, building this with `balena push` results in this error:

```
[Error]     An error occured: ERR value is not an integer or out of range
[Info]      Built on arm02
[Error]     Not deploying release.
Remote build failed
```

This was tried exactly once; it's not clear to me what's going on
here.  Obviously it needs more work, but I'm getting this branch off
my laptop.

Change-type: minor
Signed-off-by: Hugh Brown <hugh@balena.io>