Use Alpine netdata package for the container
By using the Alpine package in an Alpine container, instead of the
Netdata container, we shrink the container size from ~ 300MB to 20MB.

One note: graphs for the various containers show up approximately 1
minute after the netdata container starts up. Presumably this is the
time it takes netdata to ingest data on the various cgroups.

Change-type: minor
Signed-off-by: Hugh Brown <hugh@balena.io>